Remove AQI TODO
We now compute an instantaneous AQI score. The AQI values are meant to
be done with a 24 hour average value of the PM2.5 particle concentration
though: we could potentially further compute a 10min and 24hr avg score
based on the 10min and 24hr average values available in the "stats" key
of the Purple Air response. That can be a future improvement.